Added config options based upon ConfigParser (https://docs.python.org/3/library/configparser.html). Finished seperating out the files.
